pyproject.toml
[project]
name = "insightcanvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
]
